refactor(chaterror): extract classification subsystem
