+++ coderd/x/chatd/chatretry/chatretry.go
@@ -5,14 +5,11 @@ package chatretry
 
 import (
 	"context"
-	"errors"
-	"fmt"
-	"regexp"
-	"strconv"
-	"strings"
 	"time"
 
 	"golang.org/x/xerrors"
+
+	"github.com/coder/coder/v2/coderd/chatd/chaterror"
 )
 
 const (
@@ -31,227 +28,18 @@ const (
 	MaxAttempts = 25
 )
 
-const (
-	errorKindOverloaded     = "overloaded"
-	errorKindRateLimit      = "rate_limit"
-	errorKindTimeout        = "timeout"
-	errorKindStartupTimeout = "startup_timeout"
-	errorKindAuth           = "auth"
-	errorKindConfig         = "config"
-	errorKindGeneric        = "generic"
-)
+type ClassifiedError = chaterror.ClassifiedError
 
 var (
-	statusCodePattern       = regexp.MustCompile(`(?i)(?:status(?:\s+code)?|http)\s*[:=]?\s*(\d{3})`)
-	standaloneStatusPattern = regexp.MustCompile(`\b(?:401|403|408|429|500|502|503|504|529)\b`)
+	Classify           = chaterror.Classify
+	WithProvider       = chaterror.WithProvider
+	WithClassification = chaterror.WithClassification
 )
-
-var overloadedPatterns = []string{
-	"overloaded",
-	"overloaded_error",
-}
-
-var rateLimitPatterns = []string{
-	"rate limit",
-	"rate_limit",
-	"rate limited",
-	"rate-limited",
-	"too many requests",
-}
-
-var timeoutPatterns = []string{
-	"timeout",
-	"timed out",
-	"service unavailable",
-	"unavailable",
-	"connection reset",
-	"connection refused",
-	"eof",
-	"broken pipe",
-	"bad gateway",
-	"gateway timeout",
-}
-
-var authPatterns = []string{
-	"authentication",
-	"unauthorized",
-	"forbidden",
-	"invalid api key",
-	"invalid_api_key",
-	"quota",
-	"billing",
-	"insufficient_quota",
-	"payment required",
-}
-
-var configPatterns = []string{
-	"invalid model",
-	"model not found",
-	"model_not_found",
-	"unsupported model",
-	"context length exceeded",
-	"context_exceeded",
-	"maximum context length",
-	"malformed config",
-	"malformed configuration",
-}
-
-var genericRetryablePatterns = []string{
-	"server error",
-	"internal server error",
-}
-
-// ClassifiedError is the normalized, user-facing view of an
-// underlying provider or runtime error.
-type ClassifiedError struct {
-	Message    string
-	Kind       string
-	Provider   string
-	Retryable  bool
-	StatusCode int
-}
-
-// WithProvider returns a copy of the classification with Provider
-// filled when the classifier could not infer one from the error text.
-func (c ClassifiedError) WithProvider(provider string) ClassifiedError {
-	if strings.TrimSpace(c.Provider) != "" {
-		return c
-	}
-	c.Provider = normalizeProvider(provider)
-	c.Message = ""
-	return normalizeClassification(c)
-}
-
-// WithClassification wraps err so future calls to ClassifyError return
-// classified instead of re-deriving it from err.Error().
-func WithClassification(err error, classified ClassifiedError) error {
-	if err == nil {
-		return nil
-	}
-	return &classifiedError{
-		cause:      err,
-		classified: normalizeClassification(classified),
-	}
-}
-
-type classifiedError struct {
-	cause      error
-	classified ClassifiedError
-}
-
-func (e *classifiedError) Error() string {
-	if e == nil || e.cause == nil {
-		return ""
-	}
-	return e.cause.Error()
-}
-
-func (e *classifiedError) Unwrap() error {
-	if e == nil {
-		return nil
-	}
-	return e.cause
-}
-
-// ClassifyError normalizes err into a stable, user-facing payload used
-// for retry handling, streamed terminal errors, and persisted last_error
-// values.
-func ClassifyError(err error) ClassifiedError {
-	if err == nil {
-		return ClassifiedError{}
-	}
-
-	var wrapped *classifiedError
-	if errors.As(err, &wrapped) && wrapped != nil {
-		return normalizeClassification(wrapped.classified)
-	}
-
-	message := strings.TrimSpace(err.Error())
-	if message == "" {
-		return ClassifiedError{}
-	}
-
-	lower := strings.ToLower(message)
-	statusCode := extractStatusCode(lower)
-	provider := extractProvider(lower, statusCode)
-
-	if errors.Is(err, context.Canceled) || strings.Contains(lower, "context canceled") {
-		return normalizeClassification(ClassifiedError{
-			Message:    "The request was canceled before it completed.",
-			Kind:       errorKindGeneric,
-			Provider:   provider,
-			Retryable:  false,
-			StatusCode: statusCode,
-		})
-	}
-
-	if errors.Is(err, context.DeadlineExceeded) ||
-		strings.Contains(lower, "context deadline exceeded") {
-		return normalizeClassification(ClassifiedError{
-			Kind:       errorKindTimeout,
-			Provider:   provider,
-			Retryable:  false,
-			StatusCode: statusCode,
-		})
-	}
-
-	switch {
-	case isOverloaded(lower, statusCode):
-		return normalizeClassification(ClassifiedError{
-			Kind:       errorKindOverloaded,
-			Provider:   provider,
-			Retryable:  true,
-			StatusCode: statusCode,
-		})
-	case isConfig(lower):
-		return normalizeClassification(ClassifiedError{
-			Kind:       errorKindConfig,
-			Provider:   provider,
-			Retryable:  false,
-			StatusCode: statusCode,
-		})
-	case isAuth(lower, statusCode):
-		return normalizeClassification(ClassifiedError{
-			Kind:       errorKindAuth,
-			Provider:   provider,
-			Retryable:  false,
-			StatusCode: statusCode,
-		})
-	case isRateLimit(lower, statusCode):
-		return normalizeClassification(ClassifiedError{
-			Kind:       errorKindRateLimit,
-			Provider:   provider,
-			Retryable:  true,
-			StatusCode: statusCode,
-		})
-	case isTimeout(lower, statusCode):
-		return normalizeClassification(ClassifiedError{
-			Kind:       errorKindTimeout,
-			Provider:   provider,
-			Retryable:  true,
-			StatusCode: statusCode,
-		})
-	case isGenericRetryable(lower, statusCode):
-		return normalizeClassification(ClassifiedError{
-			Kind:       errorKindGeneric,
-			Provider:   provider,
-			Retryable:  true,
-			StatusCode: statusCode,
-		})
-	default:
-		return normalizeClassification(ClassifiedError{
-			Kind:       errorKindGeneric,
-			Provider:   provider,
-			Retryable:  false,
-			StatusCode: statusCode,
-		})
-	}
-}
 
 // IsRetryable determines whether an error from an LLM provider is
 // transient and worth retrying.
 func IsRetryable(err error) bool {
-	return ClassifyError(err).Retryable
+	return chaterror.Classify(err).Retryable
 }
 
 // StatusCodeRetryable returns true for HTTP status codes that
@@ -304,9 +92,9 @@ func Retry(ctx context.Context, fn RetryFn, onRetry OnRetryFn) error {
 			return nil
 		}
 
-		classified := ClassifyError(err)
+		classified := chaterror.Classify(err)
 		if !classified.Retryable {
-			return WithClassification(err, classified)
+			return chaterror.WithClassification(err, classified)
 		}
 
 		// If the caller's context is already done, return the
@@ -317,7 +105,7 @@ func Retry(ctx context.Context, fn RetryFn, onRetry OnRetryFn) error {
 
 		attempt++
 		if attempt >= MaxAttempts {
-			return WithClassification(
+			return chaterror.WithClassification(
 				xerrors.Errorf("max retry attempts (%d) exceeded: %w", MaxAttempts, err),
 				classified,
 			)
@@ -337,273 +125,4 @@ func Retry(ctx context.Context, fn RetryFn, onRetry OnRetryFn) error {
 		case <-timer.C:
 		}
 	}
-}
-
-func normalizeClassification(classified ClassifiedError) ClassifiedError {
-	classified.Message = strings.TrimSpace(classified.Message)
-	classified.Kind = strings.TrimSpace(classified.Kind)
-	classified.Provider = normalizeProvider(classified.Provider)
-	if classified.Kind == "" && classified.Message == "" {
-		return ClassifiedError{}
-	}
-	if classified.Kind == "" {
-		classified.Kind = errorKindGeneric
-	}
-	if classified.Message == "" {
-		classified.Message = userFacingMessage(classified)
-	}
-	return classified
-}
-
-func userFacingMessage(classified ClassifiedError) string {
-	switch classified.Kind {
-	case errorKindOverloaded:
-		return overloadedMessage(classified.Provider, classified.StatusCode)
-	case errorKindRateLimit:
-		return rateLimitMessage(classified.Provider, classified.StatusCode)
-	case errorKindTimeout:
-		return timeoutMessage(classified)
-	case errorKindStartupTimeout:
-		return startupTimeoutMessage(classified.Provider)
-	case errorKindAuth:
-		return authMessage(classified.Provider)
-	case errorKindConfig:
-		return configMessage(classified.Provider)
-	default:
-		return genericMessage(classified)
-	}
-}
-
-func overloadedMessage(provider string, statusCode int) string {
-	subject := providerSubject(provider)
-	if statusCode > 0 {
-		return fmt.Sprintf(
-			"%s is temporarily overloaded (HTTP %d). Please try again later.",
-			subject,
-			statusCode,
-		)
-	}
-	return fmt.Sprintf("%s is temporarily overloaded. Please try again later.", subject)
-}
-
-func rateLimitMessage(provider string, statusCode int) string {
-	subject := providerSubject(provider)
-	if statusCode > 0 {
-		return fmt.Sprintf(
-			"%s is rate limiting requests (HTTP %d). Please try again later.",
-			subject,
-			statusCode,
-		)
-	}
-	return fmt.Sprintf("%s is rate limiting requests. Please try again later.", subject)
-}
-
-func timeoutMessage(classified ClassifiedError) string {
-	subject := providerSubject(classified.Provider)
-	if classified.StatusCode > 0 {
-		return fmt.Sprintf(
-			"%s is temporarily unavailable (HTTP %d). Please try again later.",
-			subject,
-			classified.StatusCode,
-		)
-	}
-	if classified.Retryable {
-		return fmt.Sprintf("%s did not respond in time. Please try again.", subject)
-	}
-	return "The request timed out before it completed. Please try again."
-}
-
-func startupTimeoutMessage(provider string) string {
-	return fmt.Sprintf(
-		"%s did not start responding in time. Please try again.",
-		providerSubject(provider),
-	)
-}
-
-func authMessage(provider string) string {
-	if displayName := providerDisplayName(provider); displayName != "" {
-		return fmt.Sprintf(
-			"Authentication with %s failed. Check the API key, permissions, and billing settings.",
-			displayName,
-		)
-	}
-	return "Authentication with the AI provider failed. Check the API key, permissions, and billing settings."
-}
-
-func configMessage(provider string) string {
-	subject := providerSubject(provider)
-	return fmt.Sprintf(
-		"%s rejected the model configuration. Check the selected model and provider settings.",
-		subject,
-	)
-}
-
-func genericMessage(classified ClassifiedError) string {
-	subject := providerSubject(classified.Provider)
-	if classified.StatusCode > 0 {
-		message := fmt.Sprintf(
-			"%s returned an unexpected error (HTTP %d).",
-			subject,
-			classified.StatusCode,
-		)
-		if classified.Retryable {
-			return message + " Please try again later."
-		}
-		return message + " Please try again."
-	}
-	if classified.Retryable {
-		return fmt.Sprintf(
-			"%s returned an unexpected error. Please try again later.",
-			subject,
-		)
-	}
-	return "The chat request failed unexpectedly. Please try again."
-}
-
-func providerSubject(provider string) string {
-	if displayName := providerDisplayName(provider); displayName != "" {
-		return displayName
-	}
-	return "The AI provider"
-}
-
-func providerDisplayName(provider string) string {
-	switch normalizeProvider(provider) {
-	case "anthropic":
-		return "Anthropic"
-	case "azure":
-		return "Azure OpenAI"
-	case "bedrock":
-		return "AWS Bedrock"
-	case "google":
-		return "Google"
-	case "openai":
-		return "OpenAI"
-	case "openai-compat":
-		return "OpenAI Compatible"
-	case "openrouter":
-		return "OpenRouter"
-	case "vercel":
-		return "Vercel AI Gateway"
-	default:
-		return ""
-	}
-}
-
-func normalizeProvider(provider string) string {
-	normalized := strings.ToLower(strings.TrimSpace(provider))
-	switch normalized {
-	case "azure openai", "azure-openai":
-		return "azure"
-	case "openai compat", "openai compatible", "openai_compat":
-		return "openai-compat"
-	default:
-		return normalized
-	}
-}
-
-func extractStatusCode(lower string) int {
-	if matches := statusCodePattern.FindStringSubmatch(lower); len(matches) == 2 {
-		return atoi(matches[1])
-	}
-	if match := standaloneStatusPattern.FindString(lower); match != "" {
-		return atoi(match)
-	}
-	switch {
-	case containsAny(lower, "too many requests"):
-		return 429
-	case containsAny(lower, "unauthorized", "invalid api key", "invalid_api_key"):
-		return 401
-	case containsAny(lower, "forbidden"):
-		return 403
-	case containsAny(lower, "gateway timeout"):
-		return 504
-	case containsAny(lower, "bad gateway"):
-		return 502
-	case containsAny(lower, "service unavailable"):
-		return 503
-	case containsAny(lower, "overloaded"):
-		return 529
-	default:
-		return 0
-	}
-}
-
-func extractProvider(lower string, statusCode int) string {
-	switch {
-	case containsAny(lower, "openai-compat", "openai compatible"):
-		return "openai-compat"
-	case containsAny(lower, "azure openai", "azure-openai"):
-		return "azure"
-	case containsAny(lower, "openrouter"):
-		return "openrouter"
-	case containsAny(lower, "aws bedrock", "bedrock"):
-		return "bedrock"
-	case containsAny(lower, "vercel ai gateway", "vercel"):
-		return "vercel"
-	case containsAny(lower, "anthropic", "claude"):
-		return "anthropic"
-	case containsAny(lower, "google", "gemini", "vertex"):
-		return "google"
-	case containsAny(lower, "openai"):
-		return "openai"
-	case statusCode == 529 || containsAny(lower, "overloaded", "overloaded_error"):
-		return "anthropic"
-	default:
-		return ""
-	}
-}
-
-func isOverloaded(lower string, statusCode int) bool {
-	return statusCode == 529 || containsAny(lower, overloadedPatterns...)
-}
-
-func isRateLimit(lower string, statusCode int) bool {
-	if containsAny(lower, "quota", "billing", "insufficient_quota") {
-		return false
-	}
-	return statusCode == 429 || containsAny(lower, rateLimitPatterns...)
-}
-
-func isTimeout(lower string, statusCode int) bool {
-	switch statusCode {
-	case 408, 502, 503, 504:
-		return true
-	default:
-		return containsAny(lower, timeoutPatterns...)
-	}
-}
-
-func isAuth(lower string, statusCode int) bool {
-	switch statusCode {
-	case 401, 403:
-		return true
-	default:
-		return containsAny(lower, authPatterns...)
-	}
-}
-
-func isConfig(lower string) bool {
-	return containsAny(lower, configPatterns...)
-}
-
-func isGenericRetryable(lower string, statusCode int) bool {
-	return statusCode == 500 || containsAny(lower, genericRetryablePatterns...)
-}
-
-func containsAny(lower string, patterns ...string) bool {
-	for _, pattern := range patterns {
-		if strings.Contains(lower, pattern) {
-			return true
-		}
-	}
-	return false
-}
-
-func atoi(value string) int {
-	parsed, err := strconv.Atoi(value)
-	if err != nil {
-		return 0
-	}
-	return parsed
 }
