+++ coderd/x/chatd/chatretry/chatretry_test.go
@@ -11,106 +11,11 @@ import (
 	"github.com/stretchr/testify/require"
 	"golang.org/x/xerrors"
 
+	"github.com/coder/coder/v2/coderd/chatd/chaterror"
 	"github.com/coder/coder/v2/coderd/x/chatd/chatretry"
 )
 
-func TestClassifyError(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name       string
-		err        error
-		message    string
-		kind       string
-		provider   string
-		retryable  bool
-		statusCode int
-	}{
-		{
-			name:       "AnthropicOverloaded",
-			err:        xerrors.New("anthropic API error: status 529 overloaded_error: Overloaded"),
-			message:    "Anthropic is temporarily overloaded (HTTP 529). Please try again later.",
-			kind:       "overloaded",
-			provider:   "anthropic",
-			retryable:  true,
-			statusCode: 529,
-		},
-		{
-			name:       "RateLimitByStatusCode",
-			err:        xerrors.New("received status 429 from upstream"),
-			message:    "The AI provider is rate limiting requests (HTTP 429). Please try again later.",
-			kind:       "rate_limit",
-			provider:   "",
-			retryable:  true,
-			statusCode: 429,
-		},
-		{
-			name:       "Timeout",
-			err:        xerrors.New("dial tcp: i/o timeout"),
-			message:    "The AI provider did not respond in time. Please try again.",
-			kind:       "timeout",
-			provider:   "",
-			retryable:  true,
-			statusCode: 0,
-		},
-		{
-			name:       "Auth",
-			err:        xerrors.New("invalid api key"),
-			message:    "Authentication with the AI provider failed. Check the API key, permissions, and billing settings.",
-			kind:       "auth",
-			provider:   "",
-			retryable:  false,
-			statusCode: 401,
-		},
-		{
-			name:       "Config",
-			err:        xerrors.New("model not found"),
-			message:    "The AI provider rejected the model configuration. Check the selected model and provider settings.",
-			kind:       "config",
-			provider:   "",
-			retryable:  false,
-			statusCode: 0,
-		},
-		{
-			name:       "Generic",
-			err:        xerrors.New("boom"),
-			message:    "The chat request failed unexpectedly. Please try again.",
-			kind:       "generic",
-			provider:   "",
-			retryable:  false,
-			statusCode: 0,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			classified := chatretry.ClassifyError(tt.err)
-			require.Equal(t, tt.message, classified.Message)
-			require.Equal(t, tt.kind, classified.Kind)
-			require.Equal(t, tt.provider, classified.Provider)
-			require.Equal(t, tt.retryable, classified.Retryable)
-			require.Equal(t, tt.statusCode, classified.StatusCode)
-		})
-	}
-}
-
-func TestClassifyError_UsesWrappedClassification(t *testing.T) {
-	t.Parallel()
-
-	classified := chatretry.ClassifyError(
-		xerrors.New("received status 429 from upstream"),
-	).WithProvider("openai")
-	wrapped := chatretry.WithClassification(
-		xerrors.New("max retry attempts exceeded: received status 429 from upstream"),
-		classified,
-	)
-
-	require.Equal(t, classified, chatretry.ClassifyError(wrapped))
-}
-
-func TestIsRetryable(t *testing.T) {
+func TestIsRetryableDelegatesToClassification(t *testing.T) {
 	t.Parallel()
 
 	tests := []struct {
@@ -118,214 +23,19 @@ func TestIsRetryable(t *testing.T) {
 		err       error
 		retryable bool
 	}{
-		// Retryable errors.
-		{
-			name:      "Overloaded",
-			err:       xerrors.New("model is overloaded, please try again"),
-			retryable: true,
-		},
-		{
-			name:      "RateLimit",
-			err:       xerrors.New("rate limit exceeded"),
-			retryable: true,
-		},
-		{
-			name:      "RateLimitUnderscore",
-			err:       xerrors.New("rate_limit: too many requests"),
-			retryable: true,
-		},
-		{
-			name:      "TooManyRequests",
-			err:       xerrors.New("too many requests"),
-			retryable: true,
-		},
-		{
-			name:      "HTTP429InMessage",
-			err:       xerrors.New("received status 429 from upstream"),
-			retryable: true,
-		},
-		{
-			name:      "HTTP529InMessage",
-			err:       xerrors.New("received status 529 from upstream"),
-			retryable: true,
-		},
-		{
-			name:      "ServerError500",
-			err:       xerrors.New("status 500: internal server error"),
-			retryable: true,
-		},
-		{
-			name:      "ServerErrorGeneric",
-			err:       xerrors.New("server error"),
-			retryable: true,
-		},
-		{
-			name:      "ConnectionReset",
-			err:       xerrors.New("read tcp: connection reset by peer"),
-			retryable: true,
-		},
-		{
-			name:      "ConnectionRefused",
-			err:       xerrors.New("dial tcp: connection refused"),
-			retryable: true,
-		},
-		{
-			name:      "EOF",
-			err:       xerrors.New("unexpected EOF"),
-			retryable: true,
-		},
-		{
-			name:      "BrokenPipe",
-			err:       xerrors.New("write: broken pipe"),
-			retryable: true,
-		},
-		{
-			name:      "NetworkTimeout",
-			err:       xerrors.New("i/o timeout"),
-			retryable: true,
-		},
-		{
-			name:      "ServiceUnavailable",
-			err:       xerrors.New("service unavailable"),
-			retryable: true,
-		},
-		{
-			name:      "Unavailable",
-			err:       xerrors.New("the service is currently unavailable"),
-			retryable: true,
-		},
-		{
-			name:      "Status502",
-			err:       xerrors.New("status 502: bad gateway"),
-			retryable: true,
-		},
-		{
-			name:      "Status503",
-			err:       xerrors.New("status 503"),
-			retryable: true,
-		},
-
-		// Non-retryable errors.
-		{
-			name:      "Nil",
-			err:       nil,
-			retryable: false,
-		},
-		{
-			name:      "ContextCanceled",
-			err:       context.Canceled,
-			retryable: false,
-		},
-		{
-			name:      "ContextCanceledWrapped",
-			err:       xerrors.Errorf("operation failed: %w", context.Canceled),
-			retryable: false,
-		},
-		{
-			name:      "ContextCanceledMessage",
-			err:       xerrors.New("context canceled"),
-			retryable: false,
-		},
-		{
-			name:      "ContextDeadlineExceeded",
-			err:       xerrors.New("context deadline exceeded"),
-			retryable: false,
-		},
-		{
-			name:      "Authentication",
-			err:       xerrors.New("authentication failed"),
-			retryable: false,
-		},
-		{
-			name:      "Unauthorized",
-			err:       xerrors.New("401 Unauthorized"),
-			retryable: false,
-		},
-		{
-			name:      "Forbidden",
-			err:       xerrors.New("403 Forbidden"),
-			retryable: false,
-		},
-		{
-			name:      "InvalidAPIKey",
-			err:       xerrors.New("invalid api key"),
-			retryable: false,
-		},
-		{
-			name:      "InvalidAPIKeyUnderscore",
-			err:       xerrors.New("invalid_api_key"),
-			retryable: false,
-		},
-		{
-			name:      "InvalidModel",
-			err:       xerrors.New("invalid model: gpt-5-turbo"),
-			retryable: false,
-		},
-		{
-			name:      "ModelNotFound",
-			err:       xerrors.New("model not found"),
-			retryable: false,
-		},
-		{
-			name:      "ModelNotFoundUnderscore",
-			err:       xerrors.New("model_not_found"),
-			retryable: false,
-		},
-		{
-			name:      "ContextLengthExceeded",
-			err:       xerrors.New("context length exceeded"),
-			retryable: false,
-		},
-		{
-			name:      "ContextExceededUnderscore",
-			err:       xerrors.New("context_exceeded"),
-			retryable: false,
-		},
-		{
-			name:      "MaximumContextLength",
-			err:       xerrors.New("maximum context length"),
-			retryable: false,
-		},
-		{
-			name:      "QuotaExceeded",
-			err:       xerrors.New("quota exceeded"),
-			retryable: false,
-		},
-		{
-			name:      "BillingError",
-			err:       xerrors.New("billing issue: payment required"),
-			retryable: false,
-		},
-
-		// Wrapped errors preserve retryability.
-		{
-			name:      "WrappedRetryable",
-			err:       xerrors.Errorf("provider call failed: %w", xerrors.New("service unavailable")),
-			retryable: true,
-		},
-		{
-			name:      "WrappedNonRetryable",
-			err:       xerrors.Errorf("provider call failed: %w", xerrors.New("invalid api key")),
-			retryable: false,
-		},
+		{name: "Nil", err: nil, retryable: false},
+		{name: "RetryableStatus", err: xerrors.New("received status 429 from upstream"), retryable: true},
+		{name: "RetryableTimeout", err: xerrors.New("service unavailable"), retryable: true},
+		{name: "NonRetryableAuth", err: xerrors.New("invalid api key"), retryable: false},
+		{name: "NonRetryableGeneric", err: xerrors.New("boom"), retryable: false},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
-			classified := chatretry.ClassifyError(tt.err)
-			got := chatretry.IsRetryable(tt.err)
-			if got != tt.retryable {
-				t.Errorf("IsRetryable(%v) = %v, want %v", tt.err, got, tt.retryable)
-			}
-			if classified.Retryable != got {
-				t.Errorf(
-					"ClassifyError(%v).Retryable = %v, want %v",
-					tt.err,
-					classified.Retryable,
-					got,
-				)
-			}
+
+			require.Equal(t, tt.retryable, chatretry.IsRetryable(tt.err))
+			require.Equal(t, chaterror.Classify(tt.err).Retryable, chatretry.IsRetryable(tt.err))
 		})
 	}
 }
@@ -375,8 +85,8 @@ func TestDelay(t *testing.T) {
 		{3, 8 * time.Second},
 		{4, 16 * time.Second},
 		{5, 32 * time.Second},
-		{6, 60 * time.Second},  // Capped at MaxDelay.
-		{10, 60 * time.Second}, // Still capped.
+		{6, 60 * time.Second},
+		{10, 60 * time.Second},
 		{100, 60 * time.Second},
 	}
 
@@ -399,12 +109,8 @@ func TestRetry_SuccessOnFirstTry(t *testing.T) {
 		calls++
 		return nil
 	}, nil)
-	if err != nil {
-		t.Fatalf("expected nil error, got %v", err)
-	}
-	if calls != 1 {
-		t.Fatalf("expected fn called once, got %d", calls)
-	}
+	require.NoError(t, err)
+	require.Equal(t, 1, calls)
 }
 
 func TestRetry_TransientThenSuccess(t *testing.T) {
@@ -418,12 +124,8 @@ func TestRetry_TransientThenSuccess(t *testing.T) {
 		}
 		return nil
 	}, nil)
-	if err != nil {
-		t.Fatalf("expected nil error, got %v", err)
-	}
-	if calls != 2 {
-		t.Fatalf("expected fn called twice, got %d", calls)
-	}
+	require.NoError(t, err)
+	require.Equal(t, 2, calls)
 }
 
 func TestRetry_MultipleTransientThenSuccess(t *testing.T) {
@@ -437,12 +139,8 @@ func TestRetry_MultipleTransientThenSuccess(t *testing.T) {
 		}
 		return nil
 	}, nil)
-	if err != nil {
-		t.Fatalf("expected nil error, got %v", err)
-	}
-	if calls != 4 {
-		t.Fatalf("expected fn called 4 times, got %d", calls)
-	}
+	require.NoError(t, err)
+	require.Equal(t, 4, calls)
 }
 
 func TestRetry_NonRetryableError(t *testing.T) {
@@ -454,15 +152,14 @@ func TestRetry_NonRetryableError(t *testing.T) {
 		return xerrors.New("invalid api key")
 	}, nil)
 
-	if err == nil {
-		t.Fatal("expected error, got nil")
-	}
-	if err.Error() != "invalid api key" {
-		t.Fatalf("expected 'invalid api key', got %q", err.Error())
-	}
-	if calls != 1 {
-		t.Fatalf("expected fn called once, got %d", calls)
-	}
+	require.Error(t, err)
+	require.EqualError(t, err, "invalid api key")
+	require.Equal(t, 1, calls)
+	require.Equal(
+		t,
+		chaterror.Classify(xerrors.New("invalid api key")),
+		chaterror.Classify(err),
+	)
 }
 
 func TestRetry_ContextCanceledDuringWait(t *testing.T) {
@@ -473,8 +170,6 @@ func TestRetry_ContextCanceledDuringWait(t *testing.T) {
 	calls := 0
 	err := chatretry.Retry(ctx, func(_ context.Context) error {
 		calls++
-		// Cancel after the first retryable error so the wait
-		// select picks up the cancellation.
 		if calls == 1 {
 			cancel()
 		}
@@ -493,8 +188,6 @@ func TestRetry_ContextCanceledDuringFn(t *testing.T) {
 
 	err := chatretry.Retry(ctx, func(_ context.Context) error {
 		cancel()
-		// Return a retryable error; the loop should detect that
-		// ctx is done and return the context error.
 		return xerrors.New("overloaded")
 	}, nil)
 
@@ -509,11 +202,7 @@ func TestRetry_OnRetryCalledWithCorrectArgs(t *testing.T) {
 	type retryRecord struct {
 		attempt    int
 		errMsg     string
-		message    string
-		kind       string
-		provider   string
-		retryable  bool
-		statusCode int
+		classified chatretry.ClassifiedError
 		delay      time.Duration
 	}
 	var records []retryRecord
@@ -534,29 +223,19 @@ func TestRetry_OnRetryCalledWithCorrectArgs(t *testing.T) {
 		records = append(records, retryRecord{
 			attempt:    attempt,
 			errMsg:     err.Error(),
-			message:    classified.Message,
-			kind:       classified.Kind,
-			provider:   classified.Provider,
-			retryable:  classified.Retryable,
-			statusCode: classified.StatusCode,
+			classified: classified,
 			delay:      delay,
 		})
 	})
 	require.NoError(t, err)
 	require.Len(t, records, 2)
 
+	expected := chaterror.Classify(xerrors.New("received status 429 from upstream"))
 	require.Equal(t, 1, records[0].attempt)
 	require.Equal(t, 2, records[1].attempt)
 	require.Equal(t, "received status 429 from upstream", records[0].errMsg)
-	require.Equal(
-		t,
-		"The AI provider is rate limiting requests (HTTP 429). Please try again later.",
-		records[0].message,
-	)
-	require.Equal(t, "rate_limit", records[0].kind)
-	require.Empty(t, records[0].provider)
-	require.True(t, records[0].retryable)
-	require.Equal(t, 429, records[0].statusCode)
+	require.Equal(t, expected, records[0].classified)
+	require.Equal(t, expected, records[1].classified)
 	require.Equal(t, chatretry.Delay(0), records[0].delay)
 	require.Equal(t, chatretry.Delay(1), records[1].delay)
 }
